[QACI-627] Refactor / simplify MicroProfile TCK (#145)
* refactored POMs, comsolidated arquillian.xml into one and internal environment setting

* flattened out directory structure

* Renamed artifact IDs to better suited TCK names, i.e. health-tck instead of tck-runner
per PR review suggestions (thanks Patrik)

* removed version numbers from Java code and moved it to parent POM where it belongs

* update Arquillian Connectors
