+++ MicroProfile-Config/src/test/resources/environment-variables.properties
@@ -1,0 +1,8 @@
+config_ordinal=45
+customer_name=Bob
+my_int_property=45
+MY_BOOLEAN_PROPERTY=true
+my_string_property=haha
+MY_STRING_PROPERTY=woohoo
+MP_TCK_ENV_DUMMY=dummy
+MP_CONFIG_CACHE_DURATION=0
