+++ MicroProfile-Config/src/test/resources/system-properties.properties
@@ -1,0 +1,3 @@
+config_ordinal=120
+customer.hobby=Tennis
+mp.tck.prop.dummy=dummy
