+++ MicroProfile-Config/src/test/resources/tck-suite.xml
