--- MicroProfile-Config/tck-runner/src/test/resources/arquillian.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xmlns="http://jboss.org/schema/arquillian"
-            xsi:schemaLocation="http://jboss.org/schema/arquillian
-    http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-  
-    <container qualifier="payara-micro-managed">
-        <configuration>
-            <!-- Required for DefaultConfigSourceOrdinalTest -->
-            <property name="cmdOptions">
-                -Dconfig_ordinal=120
-                -Dcustomer.hobby=Tennis
-                -Dmp.tck.prop.dummy=dummy
-            </property>
-        </configuration>
-    </container>
-
-</arquillian>
