+++ MicroProfile-Fault-Tolerance/src/test/resources/stable-no-ext-libs-tck-suite.xml
