+++ MicroProfile-Fault-Tolerance/src/test/resources/tck-suite2.0-stable.xml
