+++ MicroProfile-Fault-Tolerance/src/test/resources/tck-suite3.0-stable.xml
@@ -1,0 +1,92 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+
+<suite name="microprofile-FT-TCK" verbose="2" >
+
+    <test name="microprofile-FT 3.0 TCK">
+        <packages>
+            <package name="org.eclipse.microprofile.fault.tolerance.tck.*" />
+        </packages>
+
+        <!-- excludes -->
+        <classes>
+            <!-- Unstable -->
+            <!--
+             -->
+
+            <!-- exception is thrown but wrapped, added JUnit tests instead -->
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.FallbackMethodWildcardNegativeTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.FallbackMethodOutOfPackageTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.FallbackMethodSuperclassPrivateTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.fallbackmethod.FallbackMethodSubclassTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.illegalConfig.IncompatibleFallbackMethodTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.illegalConfig.IncompatibleFallbackMethodWithArgsTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.illegalConfig.IncompatibleFallbackPolicies">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.illegalConfig.IncompatibleFallbackTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidAsynchronousClassTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidAsynchronousMethodTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidBulkheadAsynchQueueTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidBulkheadValueTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerDelayTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureRatioPosTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureRatioNegTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureReqVol0Test">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureReqVolNegTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureSuccessNegTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidCircuitBreakerFailureSuccess0Test">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidRetryDelayDurationTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidRetryDelayTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidRetryJitterTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidRetryMaxRetriesTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+            <class name="org.eclipse.microprofile.fault.tolerance.tck.invalidParameters.InvalidTimeoutValueTest">
+                <methods><exclude name=".*"></exclude></methods>
+            </class>
+        </classes>
+    </test>
+
+</suite>
