+++ MicroProfile-JWT-Auth/tck-arquillian-extension/pom.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
    DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-  
-   Copyright (c) [2018-2019] Payara Foundation and/or its affiliates. All rights reserved.
-  
+
+   Copyright (c) [2018-2021] Payara Foundation and/or its affiliates. All rights reserved.
+
    The contents of this file are subject to the terms of either the GNU
    General Public License Version 2 only ("GPL") or the Common Development
    and Distribution License("CDDL") (collectively, the "License").  You
@@ -12,20 +12,20 @@
    https://github.com/payara/Payara/blob/master/LICENSE.txt
    See the License for the specific
    language governing permissions and limitations under the License.
-  
+
    When distributing the software, include this License Header Notice in each
    file and include the License file at glassfish/legal/LICENSE.txt.
-  
+
    GPL Classpath Exception:
    The Payara Foundation designates this particular file as subject to the "Classpath"
    exception as provided by the Payara Foundation in the GPL Version 2 section of the License
    file that accompanied this code.
-  
+
    Modifications:
    If applicable, add the following below the License Header, with the fields
    enclosed by brackets [] replaced by your own identifying information:
    "Portions Copyright [year] [name of copyright owner]"
-  
+
    Contributor(s):
    If you wish your version of this file to be governed by only the CDDL or
    only the GPL Version 2, indicate your decision by adding "[Contributor]
@@ -47,17 +47,12 @@
         <artifactId>parent</artifactId>
         <version>1.0</version>
     </parent>
-    
+
     <artifactId>tck-arquillian-extension</artifactId>
     <packaging>jar</packaging>
 
     <name>MicroProfile JWT Auth TCK Arquillian Extension</name>
 
-    <properties>
-        <!-- JSON Dependencies -->
-        <json.version>1.1.2</json.version>
-    </properties>
-    
     <dependencies>
         <!-- JAX-RS client API dependencies -->
         <dependency>
@@ -76,7 +71,6 @@
         <dependency>
             <groupId>org.glassfish</groupId>
             <artifactId>javax.json</artifactId>
-            <version>${json.version}</version>
         </dependency>
     </dependencies>
