--- MicroProfile-JWT-Auth/tck-runner/src/test/resources/arquillian.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://jboss.org/schema/arquillian"
-  xsi:schemaLocation="http://jboss.org/schema/arquillian
-    http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-    <container qualifier="payara-micro-managed">
-        <configuration>
-            <property name="randomHttpPort">false</property>
-            <property name="autoBindHttp">false</property>
-        </configuration>
-    </container>
-
-</arquillian>
