+++ MicroProfile-JWT-Auth/tck-runner/src/test/resources/full-tck-suite.xml
@@ -1,4 +1,4 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
 
 <suite name="microprofile-jwt-auth-TCK" verbose="2" configfailurepolicy="continue" >
