+++ MicroProfile-Metrics/src/test/resources/environment-variables.properties
@@ -1,0 +1,1 @@
+MP_METRICS_TAGS=tier=integration
