+++ MicroProfile-OpenAPI/.settings/org.eclipse.core.resources.prefs
