+++ MicroProfile-OpenAPI/src/test/resources/tck-suite.xml
@@ -1,0 +1,22 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+
+<suite name="microprofile-openapi-TCK" verbose="2" configfailurepolicy="continue" >
+
+    <test name="microprofile-openapi 2.0 TCK">
+        <packages>
+            <package name="org.eclipse.microprofile.openapi.tck" />
+        </packages>
+
+        <!-- Excludes -->
+        <classes>
+            <class name="org.eclipse.microprofile.openapi.tck.ModelConstructionTest">
+                <methods>
+                    <!-- Disabled as I'm not sure this behaviour is expected -->
+                    <!-- See: https://github.com/eclipse/microprofile-open-api/issues/453 -->
+                    <exclude name="openAPITest" />
+                </methods>
+            </class>
+        </classes>
+    </test>
+
+</suite>
