--- MicroProfile-OpenAPI/tck-runner/src/test/resources/tck-suite.xml
@@ -1,22 +1,0 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="microprofile-openapi-TCK" verbose="2" configfailurepolicy="continue" >
-
-    <test name="microprofile-openapi 2.0 TCK">
-        <packages>
-            <package name="org.eclipse.microprofile.openapi.tck" />
-        </packages>
-
-        <!-- Excludes -->
-        <classes>
-            <class name="org.eclipse.microprofile.openapi.tck.ModelConstructionTest">
-                <methods>
-                    <!-- Disabled as I'm not sure this behaviour is expected -->
-                    <!-- See: https://github.com/eclipse/microprofile-open-api/issues/453 -->
-                    <exclude name="openAPITest" />
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-</suite>
