+++ MicroProfile-OpenTracing/micro-post-boot-commands.txt
