+++ MicroProfile-Rest-Client/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
