+++ src/test/resources/arquillian.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xmlns="http://jboss.org/schema/arquillian"
+            xsi:schemaLocation="http://jboss.org/schema/arquillian
+    http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
+
+    <container qualifier="payara-micro-managed">
+        <configuration>
+            <property name="randomHttpPort">${micro.randomPort}</property>
+            <property name="autoBindHttp">${micro.randomPort}</property>
+            <!-- Required for MP-Config / DefaultConfigSourceOrdinalTest -->
+            <property name="cmdOptions">
+                -Dconfig_ordinal=120
+                -Dcustomer.hobby=Tennis
+                -Dmp.tck.prop.dummy=dummy
+            </property>
+        </configuration>
+    </container>
+
+    <container qualifier="payara-embedded">
+        <configuration>
+            <property name="bindHttpPort">8080</property>
+        </configuration>
+    </container>
+
+</arquillian>
